requirements.txt
faststream[kafka]
alpaca-py
python-dotenv
clickhouse-connect
scipy
numpy